Initial support for sending GPS sync data.
Trial runs show the camera sends 0x5042 every minute or so to request a
geo update.

Testing shows the NimBLE stack deadlocks if we try to send from the
notification callback handler. Thus, we set a variable and send the geo
update via the menu handling loop.

Set the GPS coordinates to Montevideo to test signed lat/long handling.
(This in-lieu of an actual GPS unit).
Hardcode the GPS reference time to Christmas day, 2024 at 12:34:56.

Correct the final characteristic write to be an indication, not a
notification.