Hook up the M5 GPS unit via the Grove UART.
Add a settings menu entry to show real-time GPS information.
GPS is active as soon as system is started.
Must be in 'Remote Control' shutter menu for GPS data to be properly
updated.

This is still a work in progress, but looks like the major pieces are
working.
Battery life is atrocious at this stage.

LED will flash slowly to indicate GPS fix is pending.
Flash speed will increase when GPS fix is acquired.
If GPS updates are not received after 60 seconds, go back to pending.