Add support for v1.1 GPS unit.
Add a setting for the GPS unit baud rate, defaults to 9600 (the old EOL)
unit.
The new (v1.1) unit runs at 115200.

Also, add the GPS time age to the 'GPS Data' output.
Update the GPS service interval to 50ms, to support the faster unit (may
need to convert this to a FreeRTOS task/queue).

Whilst here, refactor a bit of GPS handling.