Add connection timeout support. (#204)
* Add connection timeout support.

When Infinite-Reconnect is enabled, the connection timeout is reduced to
10 seconds (default is 30 seconds).
Furthermore, the system will sleep for 10 seconds before attempting
reconnection.
This reduces the connection duty cycle to 50% and should improve battery
life at the expense of connect latency.

In testing it does take up to 10 seconds for the connection to
re-establish (as expected).
Current draw does also reduce, also as expected.

* Reduce timeout/sleep to 5 seconds.

Improve the re-connect latency by reducing to 5 seconds.
Also separate timeout and sleep durations to support non-50% duty cycle.